+++ src/utils/facilities/formatFacilityOptionsToServer.spec.js
@@ -5,6 +5,7 @@ describe('utils/facilities/formatFacilityOptionsToServer', function() {
   it('provides a default object if no options passed', function() {
     const options = formatFacilityOptionsToServer();
     expect(options).to.deep.equal({
+      include_cost_centers: false,
       include_groupings: false
     });
   });
@@ -21,10 +22,37 @@ describe('utils/facilities/formatFacilityOptionsToServer', function() {
     }, {});
 
     const expectedOptions = Object.assign({}, customOptions, {
+      include_cost_centers: false,
       include_groupings: false
     });
 
     const options = formatFacilityOptionsToServer(customOptions);
+
+    expect(options).to.deep.equal(expectedOptions);
+  });
+
+  it('provides an object that includes the `include_cost_centers` that the API service can understand', function() {
+    const expectedIncludeCostCenters = faker.random.boolean();
+    const customOptions = times(
+      faker.random.number({
+        min: 1,
+        max: 10
+      })
+    ).reduce((memo) => {
+      memo[faker.hacker.adjective()] = faker.helpers.createTransaction();
+      return memo;
+    }, {});
+
+    const expectedOptions = Object.assign({}, customOptions, {
+      include_cost_centers: expectedIncludeCostCenters,
+      include_groupings: false
+    });
+
+    const options = formatFacilityOptionsToServer({
+      ...customOptions,
+      includeCostCenters: expectedIncludeCostCenters,
+      includeGroupings: false
+    });
 
     expect(options).to.deep.equal(expectedOptions);
   });
@@ -42,11 +70,13 @@ describe('utils/facilities/formatFacilityOptionsToServer', function() {
     }, {});
 
     const expectedOptions = Object.assign({}, customOptions, {
+      include_cost_centers: false,
       include_groupings: expectedIncludeGroupings
     });
 
     const options = formatFacilityOptionsToServer({
       ...customOptions,
+      include_cost_centers: false,
       includeGroupings: expectedIncludeGroupings
     });